+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
-## [Unreleased]
+## [1.6.0-x.0] - 2019-04-26
+### Added
+- `DELETE` method for ApiCaller
+
+### Fixed
+- Parsing transaction errors without operation details
+- Wallet verification request payload
+
+## [1.5.0] - 2019-04-19
+
+## [1.5.0-rc.1] - 2019-04-19
+### Added
+- `getRaw` method & `networkDetails` getter to API caller
+
+## [1.5.0-rc.0] - 2019-04-19
 ### Added
 - `BLOB_TYPES` constant
 
@@ -59,7 +73,11 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.4.4...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.6.0-x.0...HEAD
+[1.6.0-x.0]: https://github.com/tokend/new-js-sdk/compare/1.5.0...1.6.0-x.0
+[1.5.0]: https://github.com/tokend/new-js-sdk/compare/1.5.0-rc.1...1.5.0
+[1.5.0-rc.1]: https://github.com/tokend/new-js-sdk/compare/1.5.0-rc.0...1.5.0-rc.1
+[1.5.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.4.4...1.5.0-rc.0
 [1.4.4]: https://github.com/tokend/new-js-sdk/compare/1.4.3...1.4.4
 [1.4.3]: https://github.com/tokend/new-js-sdk/compare/1.4.2...1.4.3
 [1.4.2]: https://github.com/tokend/new-js-sdk/compare/1.4.1...1.4.2